requirements.txt
argparse==1.4.0
Pillow==9.0.1